requirements.txt
# Python 3.10+ required

# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.45.0,<5.0.0
accelerate>=0.27.0
bitsandbytes>=0.41.0  # 4-bit quantisation on CUDA

# Image processing
Pillow>=9.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0  # system memory monitoring
